backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
llama-cpp-python>=0.2.0
python-multipart>=0.0.6
